refactor: コンポーネント用にindexを作成
