+++ app/.server/loaders/index.ts
@@ -1,0 +1,3 @@
+import { getAvatarRanking } from './getAvatarRanking'
+
+export { getAvatarRanking }
