+++ app/components/card/index.ts
@@ -1,0 +1,3 @@
+import { RankingCard } from './ranking-card'
+
+export { RankingCard }
