+++ app/components/layout/footer.tsx
@@ -1,0 +1,7 @@
+export const Footer = () => {
+	return (
+		<footer className="flex items-center justify-between px-6 py-9">
+			footer
+		</footer>
+	)
+}
