+++ app/components/layout/index.ts
@@ -1,0 +1,4 @@
+import { Footer } from './footer'
+import { Header } from './header'
+
+export { Footer, Header }
