+++ app/components/menu/index.ts
@@ -1,0 +1,5 @@
+import { HamburgerMenu } from './hamburger-menu'
+import { RegisterMenu } from './register-menu'
+import { UserMenu } from './user-menu'
+
+export { HamburgerMenu, RegisterMenu, UserMenu }
