+++ app/root.tsx
@@ -5,7 +5,7 @@ import {
 	Scripts,
 	ScrollRestoration,
 } from '@remix-run/react'
-import { Header } from '~/components/layout/header'
+import { Footer, Header } from '~/components/layout/index'
 import './tailwind.css'
 
 export function Layout({ children }: { children: React.ReactNode }) {
@@ -22,6 +22,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
 				<div className="flex max-w-[640px] items-center justify-center m-auto">
 					{children}
 				</div>
+				<Footer />
 				<ScrollRestoration />
 				<Scripts />
 			</body>
